+++ bin/faucet/src/main.rs
@@ -21,7 +21,7 @@ use miden_client::crypto::{RpoRandomCoin, SecretKey};
 use miden_client::store::sqlite_store::SqliteStore;
 use miden_client::{Felt, Word};
 use miden_faucet_lib::Faucet;
-use miden_faucet_lib::types::AssetOptions;
+use miden_faucet_lib::types::{AssetAmount, AssetOptions};
 use rand::{Rng, SeedableRng};
 use rand_chacha::ChaCha20Rng;
 use tokio::sync::mpsc;
@@ -44,6 +44,7 @@ const ENV_NODE_URL: &str = "MIDEN_FAUCET_NODE_URL";
 const ENV_TIMEOUT: &str = "MIDEN_FAUCET_TIMEOUT";
 const ENV_ACCOUNT_PATH: &str = "MIDEN_FAUCET_ACCOUNT_PATH";
 const ENV_ASSET_AMOUNTS: &str = "MIDEN_FAUCET_ASSET_AMOUNTS";
+const ENV_MAX_CLAIMABLE_AMOUNT: &str = "MIDEN_FAUCET_MAX_CLAIMABLE_AMOUNT";
 const ENV_REMOTE_TX_PROVER_URL: &str = "MIDEN_FAUCET_REMOTE_TX_PROVER_URL";
 const ENV_POW_SECRET: &str = "MIDEN_FAUCET_POW_SECRET";
 const ENV_POW_CHALLENGE_LIFETIME: &str = "MIDEN_FAUCET_POW_CHALLENGE_LIFETIME";
@@ -91,9 +92,13 @@ pub enum Command {
         #[arg(long = "account", value_name = "FILE", env = ENV_ACCOUNT_PATH)]
         faucet_account_path: PathBuf,
 
-        /// Comma-separated list of amounts of asset that should be dispersed on each request.
+        /// Comma-separated list of amounts of assets options to show on the frontend interface.
         #[arg(long = "asset-amounts", value_name = "U64", env = ENV_ASSET_AMOUNTS, num_args = 1.., value_delimiter = ',', default_value = "100,500,1000")]
         asset_amounts: Vec<u64>,
+
+        /// The maximum amount of assets that can be dispersed on each request.
+        #[arg(long = "max-claimable-amount", value_name = "U64", env = ENV_MAX_CLAIMABLE_AMOUNT, default_value = "1000")]
+        max_claimable_amount: u64,
 
         /// Endpoint of the remote transaction prover in the format `<protocol>://<host>[:<port>]`.
         #[arg(long = "remote-tx-prover-url", value_name = "URL", env = ENV_REMOTE_TX_PROVER_URL)]
@@ -202,6 +207,7 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
             faucet_account_path,
             remote_tx_prover_url,
             asset_amounts,
+            max_claimable_amount,
             pow_secret,
             pow_challenge_lifetime,
             pow_cleanup_interval,
@@ -242,8 +248,9 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
                 .map(|k| ApiKey::decode(k))
                 .collect::<Result<Vec<_>, _>>()
                 .context("failed to decode API keys")?;
-            let asset_options = AssetOptions::new(asset_amounts)
-                .map_err(|e| anyhow::anyhow!("failed to create asset options: {}", e))?;
+            let asset_options =
+                AssetOptions::new(asset_amounts).context("failed to create asset options")?;
+            let max_claimable_amount = AssetAmount::new(max_claimable_amount)?;
             let pow_config = PoWConfig {
                 challenge_lifetime: pow_challenge_lifetime,
                 cleanup_interval: pow_cleanup_interval,
@@ -255,6 +262,7 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
                 max_supply,
                 faucet.issuance(),
                 asset_options,
+                max_claimable_amount,
                 tx_mint_requests,
                 pow_secret.unwrap_or_default().as_str(),
                 pow_config,
@@ -471,6 +479,7 @@ mod test {
                         node_url: stub_node_url,
                         timeout: Duration::from_millis(5000),
                         asset_amounts: vec![100, 500, 1000],
+                        max_claimable_amount: 1000,
                         api_keys: vec![],
                         pow_secret: None,
                         pow_challenge_lifetime: Duration::from_secs(30),